refactor Theme component move theme option to seperate file
